refactor(git-std): split config.rs into focused modules
Split config.rs (~826 lines) into config/ submodules:
- mod.rs: type definitions (Scheme, ScopesConfig, VersioningConfig,
  ChangelogConfig, VersionFileConfig, ProjectConfig) + discover_scopes
- load.rs: TOML parsing, file discovery, config merging helpers
- tests.rs: all unit tests (previously in config.rs)

Public API unchanged. No logic changes.
